Break apart the client application and the client library
The client was rolled together in one file, but there's really a
general purpose library and some classes that implement a specific
behavior. I've broken these out into two files to make it clear that
the application is using the library so that a reader can understand
which parts they would need to implement in order to build their own
client.